Move networking-related APIs into `net-target`
- Migrate synchronization primitives to `Platform` equivalents
- Still need to excise `Config`
